Added multiple new options
Added an option to rotate images.
Added an option to specify the output quality of images (95 by default).
Added an option to stitch multiple pictures together to make panorama pictures.
Added an option to *try* to find the best crop for panorama pictures.
Added an option to find all countours, count them and, optionally, draw them.
Added an option to *try* to find the maximum best crop for panorama pictures.
Added an option to add/subtract an image from the current one.
Added an option to specify the user agent for downloading pictures.
Added an option to specify the number columns and rows for displaying differences.